+++ components/admin/pages/form/PagesForm.js
@@ -29,11 +29,14 @@ class PagesForm extends React.Component {
   });
 
   componentDidMount() {
+    const {
+      authorization,
+    } = this.props;
     const { id } = this.state;
     // Get the pages and fill the
     // state form with its params if the id exists
     if (id) {
-      fetchPage(id)
+      fetchPage(id, authorization)
         .then((data) => {
           this.setState({
             form: this.setFormFromParams(data),