+++ pages/admin/pages/index.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+// components
+import LayoutAdminStaticPages from 'layout/admin/pages';
+
+export default function AdminStaticPages() {
+  return (<LayoutAdminStaticPages />);
+}